requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
beautifulsoup4
apscheduler
lxml
python-dotenv
